requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
markdownify>=0.14.0
python-dotenv>=1.0.0
google-genai>=1.73.0
flask>=3.1.0
gunicorn>=23.0.0
